backend/requirements.txt
Django >=4.1.5
djangorestframework >= 3.14.0
psycopg2>=2.9.4
djangorestframework-simplejwt==5.2.2
django-cors-headers